# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FilmCraft - Full filmmaking production suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/beats:
    get:
      operationId: listBeats
      tags: [beats]
      summary: List story beats for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of beats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Beat"
    post:
      operationId: createBeat
      tags: [beats]
      summary: Create a beat
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBeatBody"
      responses:
        "201":
          description: Created beat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Beat"
  /projects/{projectId}/beats/{id}:
    patch:
      operationId: updateBeat
      tags: [beats]
      summary: Update a beat
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBeatBody"
      responses:
        "200":
          description: Updated beat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Beat"
    delete:
      operationId: deleteBeat
      tags: [beats]
      summary: Delete a beat
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── MIND MAP ────────────────────────────────────────────────────────────────
...
  /projects/{projectId}/casting:
    get:
      operationId: listCastingCalls
      tags: [casting]
      summary: List casting calls for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Casting calls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CastingCall"
    post:
      operationId: createCastingCall
      tags: [casting]
      summary: Create a casting call
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCastingCallBody"
      responses:
        "201":
          description: Created casting call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CastingCall"
  /projects/{projectId}/casting/{id}:
    patch:
      operationId: updateCastingCall
      tags: [casting]
      summary: Update a casting call
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCastingCallBody"
      responses:
        "200":
          description: Updated casting call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CastingCall"
    delete:
      operationId: deleteCastingCall
      tags: [casting]
      summary: Delete a casting call
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CAST MEMBERS ────────────────────────────────────────────────────────────
  /projects/{projectId}/cast:
    get:
      operationId: listCastMembers
      tags: [cast]
      summary: List cast members for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cast members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CastMember"
    post:
      operationId: createCastMember
      tags: [cast]
      summary: Add a cast member
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCastMemberBody"
      responses:
        "201":
          description: Created cast member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CastMember"
  /projects/{projectId}/cast/{id}:
    patch:
      operationId: updateCastMember
      tags: [cast]
      summary: Update a cast member
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCastMemberBody"
      responses:
        "200":
          description: Updated cast member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CastMember"
    delete:
      operationId: deleteCastMember
      tags: [cast]
      summary: Delete a cast member
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CREW ────────────────────────────────────────────────────────────────────
...
  /projects/{projectId}/budget:
    get:
      operationId: listBudgetItems
      tags: [budget]
      summary: List budget items for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetItem"
    post:
      operationId: createBudgetItem
      tags: [budget]
      summary: Create a budget item
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetItemBody"
      responses:
        "201":
          description: Created budget item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetItem"
  /projects/{projectId}/budget/{id}:
    patch:
      operationId: updateBudgetItem
      tags: [budget]
      summary: Update a budget item
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBudgetItemBody"
      responses:
        "200":
          description: Updated budget item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetItem"
    delete:
      operationId: deleteBudgetItem
      tags: [budget]
      summary: Delete a budget item
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/budget/summary:
    get:
      operationId: getBudgetSummary
      tags: [budget]
      summary: Budget summary — totals by category, spent vs allocated
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetSummary"

  # ─── LIGHTING DIAGRAMS ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Beat:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        beatType:
          type: string
          nullable: true
          description: "opening-image | theme | catalyst | debate | break-into-two | b-story | fun-and-games | midpoint | bad-guys-close-in | all-is-lost | dark-night | break-into-three | finale | final-image | custom"
        orderIndex:
          type: integer
        pageTarget:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - projectId
        - title
        - orderIndex
        - createdAt
    CreateBeatBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        beatType:
          type: string
          nullable: true
        orderIndex:
          type: integer
        pageTarget:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - title
        - orderIndex
    UpdateBeatBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        beatType:
          type: string
          nullable: true
        orderIndex:
          type: integer
        pageTarget:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true

    # ── MIND MAP NODE ─────────────────────────────────────────────────────────
...
    CastingCall:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        characterName:
          type: string
        role:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        ageRange:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
          nullable: true
        auditionDate:
          type: string
          nullable: true
        auditionLocation:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
          description: "open | in-review | cast | closed"
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - projectId
        - characterName
        - createdAt
    CreateCastingCallBody:
      type: object
      properties:
        characterName:
          type: string
        role:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        ageRange:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
          nullable: true
        auditionDate:
          type: string
          nullable: true
        auditionLocation:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - characterName
    UpdateCastingCallBody:
      type: object
      properties:
        characterName:
          type: string
        role:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        ageRange:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
          nullable: true
        auditionDate:
          type: string
          nullable: true
        auditionLocation:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true

    # ── CAST MEMBER ───────────────────────────────────────────────────────────
    CastMember:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        actorName:
          type: string
        characterName:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        contact:
          type: string
          nullable: true
        agent:
          type: string
          nullable: true
        rate:
          type: number
          nullable: true
        availability:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - projectId
        - actorName
        - createdAt
    CreateCastMemberBody:
      type: object
      properties:
        actorName:
          type: string
        characterName:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        contact:
          type: string
          nullable: true
        agent:
          type: string
          nullable: true
        rate:
          type: number
          nullable: true
        availability:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - actorName
    UpdateCastMemberBody:
      type: object
      properties:
        actorName:
          type: string
        characterName:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        contact:
          type: string
          nullable: true
        agent:
          type: string
          nullable: true
        rate:
          type: number
          nullable: true
        availability:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true

    # ── CREW MEMBER ───────────────────────────────────────────────────────────
...
    BudgetItem:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        category:
          type: string
          description: "above-the-line | below-the-line | post-production | marketing | contingency | other"
        subcategory:
          type: string
          nullable: true
        description:
          type: string
        estimatedAmount:
          type: number
        actualAmount:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - projectId
        - category
        - description
        - estimatedAmount
        - createdAt
    CreateBudgetItemBody:
      type: object
      properties:
        category:
          type: string
        subcategory:
          type: string
          nullable: true
        description:
          type: string
        estimatedAmount:
          type: number
        actualAmount:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - category
        - description
        - estimatedAmount
    UpdateBudgetItemBody:
      type: object
      properties:
        category:
          type: string
        subcategory:
          type: string
          nullable: true
        description:
          type: string
        estimatedAmount:
          type: number
        actualAmount:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
    BudgetSummary:
      type: object
      properties:
        totalEstimated:
          type: number
        totalActual:
          type: number
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/BudgetCategoryTotal"
        percentSpent:
          type: number
      required:
        - totalEstimated
        - totalActual
        - byCategory
        - percentSpent
    BudgetCategoryTotal:
      type: object
      properties:
        category:
          type: string
        estimated:
          type: number
        actual:
          type: number
      required:
        - category
        - estimated
        - actual

    # ── LIGHTING DIAGRAM ──────────────────────────────────────────────────────
...